Decouple training and tuning
This lets us run without using Tune. To use tune, `--tune` must be passed on
the CLI explicitly.
