+++ safe_grid_agents/parsing/core_parser_configs.yaml
@@ -39,3 +39,6 @@ core:
     alias: dc
     action: store_true
     help: "Disable CUDA computations (default: False)"
+  tune:
+    action: store_true
+    help: "Use Ray Tune (default: False)"
